requirements.txt
aioredis==1.3.1
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
autopep8==1.7.0
billiard==3.6.4.0
black==22.3.0
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
daphne==3.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==2.2.5
Django==4.0.4
django-allauth==0.51.0
django-ckeditor==6.4.1
django-cors-headers==3.13.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-debug-toolbar==3.4.0
django-filter==21.1
django-js-asset==2.0.0
django-range-slider==0.2.4
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-writable-nested==0.6.3
drf-yasg==1.20.0
gunicorn==20.1.0
h11==0.13.0
hiredis==2.0.0
httptools==0.4.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.1
msgpack==1.0.4
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.2
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==4.3.3
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
stripe==3.4.0
toml==0.10.2
tomli==2.0.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
uvicorn==0.18.2
vine==5.0.0
watchfiles==0.15.0
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
zope.interface==5.4.0
